@@ -23,15 +23,15 @@
   "license": "MIT",
   "devDependencies": {
     "@kkt/less-modules": "7.1.1",
-    "@types/react": "17.0.39",
-    "@types/react-dom": "17.0.11",
+    "@types/react": "18.0.0",
+    "@types/react-dom": "18.0.0",
     "electron": "16.0.10",
     "kkt": "7.1.5",
     "tsbb": "3.7.0"
   },
   "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react": "18.0.0",
+    "react-dom": "18.0.0"
   },
   "eslintConfig": {
     "extends": [
