@@ -46,22 +46,22 @@
   "dependencies": {
     "@uiw/react-github-corners": "1.5.3",
     "@uiw/react-markdown-preview": "3.5.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react": "18.0.0",
+    "react-dom": "18.0.0"
   },
   "devDependencies": {
     "@kkt/less-modules": "7.1.1",
     "@kkt/raw-modules": "7.1.1",
     "@kkt/scope-plugin-options": "7.1.1",
-    "@types/react": "17.0.39",
-    "@types/react-dom": "17.0.11",
+    "@types/react": "18.0.0",
+    "@types/react-dom": "18.0.0",
     "@types/react-test-renderer": "17.0.1",
     "compile-less-cli": "1.8.11",
     "husky": "7.0.4",
     "kkt": "7.1.5",
     "lint-staged": "12.3.4",
     "prettier": "2.5.1",
-    "react-test-renderer": "17.0.2",
+    "react-test-renderer": "18.0.0",
     "tsbb": "3.7.0"
   },
   "eslintConfig": {
