@@ -15,8 +15,8 @@
   "license": "MIT",
   "dependencies": {
     "axios": "0.24.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
     "react-router-dom": "6.2.1"
   },
   "devDependencies": {
