@@ -1,11 +1,10 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactClient from 'react-dom/client';
 import './index.css';
 import App from './App';
 
-ReactDOM.render(
+ReactClient.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById('root'),
 );
