@@ -1,7 +1,7 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import App from './app/App';
 import './index.less';
 import './index.css';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+import ReactClient from 'react-dom/client';
+ReactClient.createRoot(document.getElementById('root')).render(<App />);
