@@ -1,4 +1,4 @@
-import ReactDOM from 'react-dom';
 import App from './App';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+import ReactClient from 'react-dom/client';
+ReactClient.createRoot(document.getElementById('root')!).render(<App />);
