@@ -2,8 +2,8 @@ import ReactDOM from 'react-dom';
 import App from './app/App';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
-
-ReactDOM.render(<App />, document.getElementById('root'));
+import ReactClient from 'react-dom/client';
+ReactClient.createRoot(document.getElementById('root')!).render(<App />);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
